API: rename module data.base to data.linkage
