--- src/pytools/data/base/__init__.py
@@ -1,5 +1,0 @@
-"""
-Base classes for data classes.
-"""
-
-from ._linkage import *
