+++ src/pytools/data/linkage/__init__.py
@@ -1,0 +1,5 @@
+"""
+Supporting classes for linkage trees.
+"""
+
+from ._linkage import *
