+++ src/pytools/data/linkage/_linkage.py
